fixup: refractoring multiple occurence of date validation based on month and year
