+++ src/filler.js
@@ -68,7 +68,6 @@ async function main() {
 				console.log();	
 			});
 		}
-		console.log();
 	});
 
 }
