+++ src/filler/engines/parser-engine.js
@@ -1,11 +1,14 @@
 import QType from "../../utils/question-types";
 import { isValidPhoneNumber } from "libphonenumber-js";
 import DEFAULT_TEL_COUNTRY from "../../utils/environment";
+import validationUtils from "../../utils/validation-utils";
+
+
 
 export class ParserEngine {
-    constructor() { }
-
-
+    constructor() { 
+        this.validationUtil = new validationUtils();
+    }
 
     parse(fieldType, extractedValue, response) {
         // Input Type : fieldType is a enum based on QType enum
@@ -144,29 +147,11 @@ export class ParserEngine {
         const dateArr = response === null ? false : response.match(dateValid)
         if (dateArr) {
 
-            // Months containing 31 days
-            const valid31 = ['01', '03', '05', '07', '08', '10', '12']
             const date = dateArr[1]
             const month = dateArr[2]
-            const year = Number(dateArr[3])
-
-
-            if (date == '31') {
-                if (!valid31.includes(month))
-                    return false;
-            }
-            else if (date == '30') {
-                if (month == '02') {
-                    return false; // February doesn't have 30 days
-                }
-            }
-            else if (date == '29') {
-                if (month == '02') {
-                    if (year % 4 != 0 || (year % 100 == 0 && year % 400 != 0))
-                        return false; // Non-leap year doesn't have 29 days
-                }
-            }
-            return true
+            const year = dateArr[3]
+            
+            return this.validationUtil.validateDate(date, month, year);
         }
         else {
             return false
@@ -184,29 +169,11 @@ export class ParserEngine {
         const dateArr = response === null ? false : response.match(dateTimeValid)
         if (dateArr) {
 
-            // Months containing 31 days
-            const valid31 = ['01', '03', '05', '07', '08', '10', '12']
             const date = dateArr[1]
             const month = dateArr[2]
-            const year = Number(dateArr[3])
-
-
-            if (date == '31') {
-                if (!valid31.includes(month))
-                    return false;
-            }
-            else if (date == '30') {
-                if (month == '02') {
-                    return false; // February doesn't have 30 days
-                }
-            }
-            else if (date == '29') {
-                if (month == '02') {
-                    if (year % 4 != 0 || (year % 100 == 0 && year % 400 != 0))
-                        return false; // Non-leap year doesn't have 29 days
-                }
-            }
-            return true
+            const year = dateArr[3]
+            
+            return this.validationUtil.validateDate(date, month, year);   
         }
         else
             return false
@@ -229,57 +196,42 @@ export class ParserEngine {
         // HH is in 24 hrs format
         // Checks if duration is in correct format
         // Output Type :- Boolean
+
         var durationValid = /^([01][0-9]|[2][0-3])(-[0-5][0-9]){2}$/
         return response && Boolean(response.match(durationValid))
     }
     
     validateDateWithoutYear(response) {
         // Input Type : String
-        // Valid Input format :- "dd-mm"
+        // Valid Input format :- "DD-MM"
+        // Checks if date & month is in correct format
         // Output Type : Boolean
+
         var dateWYearValid = /^([0][1-9]|[1-2][0-9]|[3][0-1])-([0][1-9]|[1][0-2])$/
         const dateArr = response === null ? false : response.match(dateWYearValid)
         if(dateArr)
         {
-            const valid31 = ['01', '03', '05', '07', '08', '10', '12']
             const date = dateArr[1]
             const month = dateArr[2]
-            if (date == '31') {
-                if (!valid31.includes(month))
-                    return false;
-            }
-            else if (date == '30') {
-                if (month == '02') {
-                    return false; // February doesn't have 30 days
-                }
-            }
-            return true;
+            return this.validationUtil.validateDate(date, month)
         }
         return false;
     }
 
     validateDateTimeWithoutYear(response) {
         // Input Type : String
-        // Valid Input format :- "dd-mm-hh-mm"
-        // hh in 24 hrs format
+        // Valid Input format :- "DD-MM-HH-mm"
+        // HH in 24 hrs format
+        // Checks if date, month, hour and minute are valid or not
         // Output Type : Boolean
+
         var dateTimeWYearValid = /^([0][1-9]|[1-2][0-9]|[3][0-1])-([0][1-9]|[1][0-2])-([01][0-9]|[2][0-3])-([0-5][0-9])$/
         const dateArr = response === null ? false : response.match(dateTimeWYearValid)
         if(dateArr)
         {
-            const valid31 = ['01', '03', '05', '07', '08', '10', '12']
             const date = dateArr[1]
             const month = dateArr[2]
-            if (date == '31') {
-                if (!valid31.includes(month))
-                    return false;
-            }
-            else if (date == '30') {
-                if (month == '02') {
-                    return false; // February doesn't have 30 days
-                }
-            }
-            return true;
+            return this.validationUtil.validateDate(date, month)
         }
         else
         {
